kickstarter_django/requirements.txt
Django==1.10.5
django-admin-tools==0.8.0
django-bootstrap3==8.1.0
django-chartjs==1.1
django-tables2==1.2.9
psycopg2==2.6.2
django-filter==1.0.1
django-cachalot==1.4.1
gunicorn==19.7.1
